pipeline/requirements.txt
youtube-transcript-api>=0.6.2
trafilatura>=1.12.0
pypdf>=4.3.0
requests>=2.32.0
python-dotenv>=1.0.1
tqdm>=4.66.0
